pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-pvoicenote-app"
version = "0.1.0"
description = "會議即時智能儀表板 — 即時轉錄 + RAG 校正 + 週期性摘要"
requires-python = ">=3.11"
dependencies = [
    "faster-whisper>=1.0.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "flet>=0.25.0",
    "ollama>=0.3.0",
    "pyyaml>=6.0",
    "sounddevice>=0.5.0",
    "pydub>=0.25.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: tests requiring ML model loading (run separately due to memory)",
]
